refactor(sync): move error into module
This is in preparation for making it a module wide error type.
