+++ crates/pathfinder/src/sync.rs
@@ -3,6 +3,7 @@
 
 mod checkpoint;
 mod class_definitions;
+mod error;
 mod headers;
 mod receipts;
 mod state_updates;
