+++ crates/pathfinder/src/sync/headers.rs
@@ -1,4 +1,5 @@
 #![allow(dead_code, unused_variables)]
+use crate::sync::error::HeaderSyncError;
 use anyhow::Context;
 use p2p::PeerData;
 use pathfinder_common::{
@@ -73,29 +74,6 @@ pub(super) async fn next_gap(
     })
     .await
     .context("Joining blocking task")?
-}
-
-#[derive(Debug, thiserror::Error)]
-pub(super) enum HeaderSyncError {
-    #[error(transparent)]
-    DatabaseError(#[from] anyhow::Error),
-    #[error("Signature verification failed")]
-    BadSignature(PeerData<SignedBlockHeader>),
-    #[error("Block hash verification failed")]
-    BadBlockHash(PeerData<SignedBlockHeader>),
-    #[error("Discontinuity in header chain")]
-    Discontinuity(PeerData<SignedBlockHeader>),
-}
-
-impl HeaderSyncError {
-    pub fn peer_id_and_data(&self) -> Option<&PeerData<SignedBlockHeader>> {
-        match self {
-            HeaderSyncError::DatabaseError(_) => None,
-            HeaderSyncError::BadSignature(x) => Some(x),
-            HeaderSyncError::BadBlockHash(x) => Some(x),
-            HeaderSyncError::Discontinuity(x) => Some(x),
-        }
-    }
 }
 
 /// Ensures the header block ID matches expectations.
